+++ src/main/java/lotto/domain/lottos/InvalidLottosException.java
@@ -1,4 +1,4 @@
-package lotto.domain.lotto;
+package lotto.domain.lottos;
 
 /**
  * Lottos 일급 컬랙션과 관련된 커스텀 예외
